src/Camera/requirements.txt
Pillow
numpy
opencv-python
paho-mqtt
pytesseract
pygame
regex
